Externalize Database to PostgreSQL
We are no longer on SQLite, and instead are targeting
PostgreSQL natively.  Since Postgres can run in a container,
and be accessible only on loopback, or on a c2c network,
this is just as good as SQLite, with better centralization
prospects.

We will not, at this time, be entertaining _other_ external
RDBMs provides (no MySQL, no Oracle, no SQL Server).

+++ db/schema_v1.go
@@ -99,9 +99,9 @@ func (s v1Schema) Deploy(db *DB) error {
 	                 target_uuid   UUID,
 
 	                 status       TEXT NOT NULL,
-	                 requested_at INTEGER NOT NULL,
-	                 started_at   INTEGER,
-	                 stopped_at   INTEGER,
+	                 requested_at BIGINT NOT NULL,
+	                 started_at   BIGINT,
+	                 stopped_at   BIGINT,
 
 	                 log       TEXT
 	               )`)
